Separate youtube and reddit functions into separate files. Trim code into data we will use
